City api Json response

+++ app/Http/Controllers/CityController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\City; 
+use App\City;
 use App\Http\Requests\CityRequest;
 use Exception;
 use Illuminate\Http\Request;
@@ -12,21 +12,22 @@ class CityController extends Controller
 {
     public function __construct()
     {
-        $this->middleware(['permission:cities_read'])->only('index');
-        $this->middleware(['permission:cities_create'])->only('create');
-        $this->middleware(['permission:cities_update'])->only('edit');
-        $this->middleware(['permission:cities_delete'])->only('destroy');
+//        $this->middleware(['permission:cities_read'])->only('index');
+//        $this->middleware(['permission:cities_create'])->only('create');
+//        $this->middleware(['permission:cities_update'])->only('edit');
+//        $this->middleware(['permission:cities_delete'])->only('destroy');
 
     }
+
     /**
      * Display a listing of the resource.
      *
      * @return Response
      */
     public function index()
-    { 
-        $cities = City::OrderBy('created_at', 'desc')->get();
-        return view('main_settings.cities.index', compact('cities'));
+    {
+        $cities = City::OrderBy('created_at', 'desc')->paginate(10);
+        return responseJson(1, "ok", $cities);
     }
 
     /**
@@ -43,51 +44,61 @@ class CityController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  Request  $request
+     * @param Request $request
      * @return Response
      */
-    public function store(CityRequest $request)
+    public function store(Request $request)
     {
+
+        $validator = validator($request->all(), [
+            "name" => "required",
+            "government_id" => 'required|exists:cities'
+        ]);
+
+        if ($validator->fails()) {
+            return responseJson(0, $validator->errors()->getMessages(), "");
+        }
         try {
             $city = City::create($request->all());
             if ($city) {
-                notfy(__('add city'), __('add city') .' '. $city->name, 'fa fa-building-o');
-                notify()->success(__('process has been success'), "", "bottomLeft" );
-                return redirect()->route('cities.index');
+                return responseJson(1, __('data created successfully'), $city);
+
             } else {
-                notify()->error("هناك خطأ ما يرجى المحاولة فى وقت لاحق", "", "bottomLeft", );
-                return redirect()->route('cities.index');
+
             }
-            
+
         } catch (Exception $th) {
-            notify()->error(" هناك خطأ ما يرجى المحاولة فى وقت لاحق", "", "bottomLeft", );
-            return redirect()->route('cities.index');
+            return responseJson(0, "", $th->getMessage());
         }
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return Response
      */
     public function show($id)
     {
-        //
+        $city = City::find($id);
+        if (!$city) {
+            return responseJson(0, __('data not found'), '');
+        }
+        return responseJson(1, "ok", $city);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return Response
      */
     public function edit($id)
     {
         $city = City::find($id);
         if (!$city) {
-            notify()->warning("هذه الاعدادات غير موجودة ", "", "bottomLeft", );
-            return redirect()->route('cities.index');
+            return responseJson(0, __('data not found'), '');
+
         }
         return view('main_settings.cities.edit', compact('city'));
     }
@@ -95,32 +106,38 @@ class CityController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  Request  $request
-     * @param  int  $id
+     * @param Request $request
+     * @param int $id
      * @return Response
      */
-    public function update(CityRequest $request, $id)
+    public function update(Request $request, $id)
     {
+        $validator = validator($request->all(), [
+            "name" => "required",
+            "government_id" => 'required|exists:cities'
+        ]);
+
+        if ($validator->fails()) {
+            return responseJson(0, $validator->errors()->getMessages(), "");
+        }
         try {
             $city = City::find($id);
             if (!$city) {
-                notify()->warning("هذه الاعدادات غير موجودة ", "", "bottomLeft", );
-                return redirect()->route('cities.index');
+                return responseJson(0, __('data not found'), '');
             } else {
                 $city->update($request->all());
-                notify()->success("تم تعديل الاعدادات بنجاح", "", "bottomLeft", );
-                return redirect()->route('cities.index');
+                return responseJson(1, __('data updated successfully'), $city);
+
             }
         } catch (Exception $ex) {
-            notify()->error(" هناك خطأ ما يرجى المحاولة فى وقت لاحق", "", "bottomLeft", );
-            return redirect()->route('cities.index');
+            return responseJson(0, "", $ex->getMessage());
         }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return Response
      */
     public function destroy($id)
@@ -128,15 +145,17 @@ class CityController extends Controller
         try {
             $city = City::find($id);
             if (!$city) {
-                notify()->warning("هذه الاعدادات غير موجودة", "", "bottomLeft", );
-                return redirect()->route('cities.index');
+                return responseJson(0, __('data not found'), '');
+            }
+
+            if ($city->applications->count() > 0 || $city->students->count()  > 0 ) {
+                return responseJson(0, __('this item can not be deleted'), $city->fresh());
             }
             $city->delete();
-            notify()->success("تم الحذف بنجاح", "", "bottomLeft", );
-            return redirect()->route('cities.index');
+            return responseJson(1, __('deleted successfully'), '');
+
         } catch (Exception $ex) {
-            notify()->error(" هناك خطأ ما يرجى المحاولة فى وقت لاحق", "", "bottomLeft", );
-            return redirect()->route('cities.index');
+            return responseJson(0, "", $ex->getMessage());
         }
     }
 }
